Update remove-runtime-sdk-versions.md
I am checking on when Visual Studio started being responsible for it's SDK removal. 

The link at the end should be to the canonical VS/SDK mapping. I reached out to Lee Coward to see if this should be in docs or download site (where it might be easier to automate). This should not delay us as we could temporarily link to the VS/SDK problem article we are working on separately.